Fixed add new process filtered
If the list is filtered and a new process is added, this , if the filter is aligned with new process, will filter also the new added proces.
For example, if are filtered all the preferred, and the new one is preferred, once added this lates process will be added to the filtered list.